pyproject.toml
[tool.poetry]
name = "hevy-api-wrapper"
version = "1.0.1"
description = "Python API wrapper for the Hevy API (sync + async, typed models, retries, pagination)"
authors = ["Dominic Kuncik <itsdomhd@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "hevy_api_wrapper", from = "src" }]
homepage = "https://github.com/dkuncik/hevy-api-wrapper"
repository = "https://github.com/dkuncik/hevy-api-wrapper"
keywords = ["hevy", "fitness", "api", "client", "httpx"]
include = ["src/hevy_api_wrapper/py.typed"]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
httpx = ">=0.27.0"
pydantic = ">=2.5.0"
typing-extensions = ">=4.8.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4"
respx = ">=0.21.1"
python-dotenv = ">=1.0.0"
pytest-asyncio = ">=0.23"
pytest-cov = ">=4.1.0"
black = ">=23.0.0"
isort = ">=5.12.0"
mypy = ">=1.7.0"
pre-commit = ">=3.5.0"

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'
skip-string-normalization = false

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["hevy_api_wrapper"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
no_lines_before = "LOCALFOLDER"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.7.0"]
build-backend = "poetry.core.masonry.api"
